Refactored bedrockify recipes and armour swap code.
