+++ src/main/java/me/juancarloscp52/bedrockify/mixin/common/features/recipes/FireworkStarFadeRecipeMixin.java
@@ -2,7 +2,7 @@ package me.juancarloscp52.bedrockify.mixin.common.features.recipes;
 
 import com.google.common.collect.Lists;
 import me.juancarloscp52.bedrockify.Bedrockify;
-import me.juancarloscp52.bedrockify.utils.DyeHelper;
+import me.juancarloscp52.bedrockify.common.features.recipes.DyeHelper;
 import net.minecraft.inventory.CraftingInventory;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
